+++ test/jsxToTag.test.js
@@ -11,25 +11,18 @@ const config = {
 
 test('<div className=\'test\'></div>', () => {
   const transformed = babel.transformSync(`<div className='test'></div>`, config);
-  expect(transformed.code).toBe(`tag("div", {
-  className: 'test'
-});`);
+  expect(transformed.code).toBe(`tag("div", 'test');`);
 });
 
 test('<div className=\'test\' id=\'mydiv\'></div>', () => {
   const transformed = babel.transformSync(`<div className='test' id='mydiv'></div>`, config);
-  expect(transformed.code).toBe(`tag("div", {
-  id: 'mydiv',
-  className: 'test'
-});`);
+  expect(transformed.code).toBe(`tag("div", 'test', 'mydiv');`);
 });
 
 test('<div className=\'test\' id=\'mydiv\' attr-section={section}></div>', () => {
   const code = `<div className='test' id='mydiv' attr-section={section}></div>`;
   const transformed = babel.transformSync(code, config);
-  expect(transformed.code).toBe(`tag("div", {
-  id: 'mydiv',
-  className: 'test',
+  expect(transformed.code).toBe(`tag("div", 'test', 'mydiv', {
   attr: {
     "section": section
   }
@@ -78,11 +71,7 @@ test(`<>...</>`, () => {
   <div className='test2'></div>
 </>`;
   const transformed = babel.transformSync(code, config);
-  expect(transformed.code).toBe(`["\\n  ", tag("div", {
-  className: 'test1'
-}), "\\n  ", tag("div", {
-  className: 'test2'
-}), "\\n"];`);
+  expect(transformed.code).toBe(`["\\n  ", tag("div", 'test1'), "\\n  ", tag("div", 'test2'), "\\n"];`);
 });
 
 test(`<Test>...</Test>`, () => {
@@ -91,35 +80,31 @@ test(`<Test>...</Test>`, () => {
   <div className='test2'></div>
 </Test>`;
   const transformed = babel.transformSync(code, config);
-  expect(transformed.code).toBe(`tag(Test, ["\\n  ", tag("div", {
-  className: 'test1'
-}), "\\n  ", tag("div", {
-  className: 'test2'
-}), "\\n"]);`);
+  expect(transformed.code).toBe(`tag(Test, ["\\n  ", tag("div", 'test1'), "\\n  ", tag("div", 'test2'), "\\n"]);`);
 });
 
 test(`<>test</>`, () => {
   const code = `const test = <>test</>`;
   const transformed = babel.transformSync(code, config);
-  expect(transformed.code).toBe(`const test = tag.use("test");`);
+  expect(transformed.code).toBe(`const test = ["test"];`);
 });
 
 test(`<>0</>`, () => {
   const code = `const test = <>0</>`;
   const transformed = babel.transformSync(code, config);
-  expect(transformed.code).toBe(`const test = tag.use(0);`);
+  expect(transformed.code).toBe(`const test = ["0"];`);
 });
 
 test(`<>{text}</>`, () => {
-  const code = `const test = <>{test}</>`;
+  const code = `const test = <>{text}</>`;
   const transformed = babel.transformSync(code, config);
-  expect(transformed.code).toBe(`const test = tag.use(test);`);
+  expect(transformed.code).toBe(`const test = [text];`);
 });
 
 test(`<></>`, () => {
   const code = `const test = <></>`;
   const transformed = babel.transformSync(code, config);
-  expect(transformed.code).toBe(`const test = tag.use();`);
+  expect(transformed.code).toBe(`const test = [];`);
 });
 
 test(`<Test {...rest} />`, () => {